backend/requirement.txt
fastapi==0.78.0
uvicorn==0.17.6
sqlalchemy
mysql-connector-python==8.0.27
python-dotenv==0.19.2
pandas

